backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
groq
python-jose[cryptography]
PyJWT
sentence-transformers
chromadb
numpy
beautifulsoup4
requests
lxml
EOF